+++ tapo-py/src/handlers/color_light_handler.rs
@@ -1,11 +1,12 @@
-use std::{ops::Deref, sync::Arc};
+use std::ops::{Deref, DerefMut};
+use std::sync::Arc;
 
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
 use tapo::requests::{Color, ColorLightSetDeviceInfoParams};
 use tapo::responses::{DeviceInfoColorLightResult, DeviceUsageEnergyMonitoringResult};
 use tapo::ColorLightHandler;
-use tokio::sync::Mutex;
+use tokio::sync::RwLock;
 
 use crate::call_handler_method;
 use crate::errors::ErrorWrapper;
@@ -14,13 +15,13 @@ use crate::runtime::tokio;
 #[derive(Clone)]
 #[pyclass(name = "ColorLightHandler")]
 pub struct PyColorLightHandler {
-    handler: Arc<Mutex<ColorLightHandler>>,
+    handler: Arc<RwLock<ColorLightHandler>>,
 }
 
 impl PyColorLightHandler {
     pub fn new(handler: ColorLightHandler) -> Self {
         Self {
-            handler: Arc::new(Mutex::new(handler)),
+            handler: Arc::new(RwLock::new(handler)),
         }
     }
 }
@@ -28,32 +29,55 @@ impl PyColorLightHandler {
 #[pymethods]
 impl PyColorLightHandler {
     pub async fn refresh_session(&self) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::refresh_session, discard_result)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.write().await.deref_mut(),
+            ColorLightHandler::refresh_session,
+            discard_result
+        )
     }
 
     pub async fn on(&self) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::on)
+        let handler = self.handler.clone();
+        call_handler_method!(handler.read().await.deref(), ColorLightHandler::on)
     }
 
     pub async fn off(&self) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::off)
+        let handler = self.handler.clone();
+        call_handler_method!(handler.read().await.deref(), ColorLightHandler::off)
     }
 
     pub async fn device_reset(&self) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::device_reset)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::device_reset
+        )
     }
 
     pub async fn get_device_info(&self) -> PyResult<DeviceInfoColorLightResult> {
-        call_handler_method!(self, ColorLightHandler::get_device_info)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::get_device_info
+        )
     }
 
     pub async fn get_device_info_json(&self) -> PyResult<Py<PyDict>> {
-        let result = call_handler_method!(self, ColorLightHandler::get_device_info_json)?;
+        let handler = self.handler.clone();
+        let result = call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::get_device_info_json,
+        )?;
         Python::with_gil(|py| tapo::python::serde_object_to_py_dict(py, &result))
     }
 
     pub async fn get_device_usage(&self) -> PyResult<DeviceUsageEnergyMonitoringResult> {
-        call_handler_method!(self, ColorLightHandler::get_device_usage)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::get_device_usage
+        )
     }
 
     pub fn set(&self) -> PyColorLightSetDeviceInfoParams {
@@ -61,20 +85,37 @@ impl PyColorLightHandler {
     }
 
     pub async fn set_brightness(&self, brightness: u8) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::set_brightness, brightness)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::set_brightness,
+            brightness
+        )
     }
 
     pub async fn set_color(&self, color: Color) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::set_color, color)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::set_color,
+            color
+        )
     }
 
     pub async fn set_hue_saturation(&self, hue: u16, saturation: u8) -> PyResult<()> {
-        call_handler_method!(self, ColorLightHandler::set_hue_saturation, hue, saturation)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.read().await.deref(),
+            ColorLightHandler::set_hue_saturation,
+            hue,
+            saturation
+        )
     }
 
     pub async fn set_color_temperature(&self, color_temperature: u16) -> PyResult<()> {
+        let handler = self.handler.clone();
         call_handler_method!(
-            self,
+            handler.read().await.deref(),
             ColorLightHandler::set_color_temperature,
             color_temperature
         )
@@ -138,7 +179,7 @@ impl PyColorLightSetDeviceInfoParams {
 
         tokio()
             .spawn(async move {
-                let handler_lock = handler.handler.lock().await;
+                let handler_lock = handler.handler.read().await;
 
                 params
                     .send(handler_lock.deref())