+++ tapo-py/src/handlers/hub_handler.rs
@@ -1,10 +1,11 @@
+use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 
 use pyo3::prelude::*;
 use pyo3::types::{PyDict, PyList};
 use tapo::responses::{ChildDeviceHubResult, DeviceInfoHubResult};
 use tapo::HubHandler;
-use tokio::sync::Mutex;
+use tokio::sync::RwLock;
 
 use crate::call_handler_method;
 use crate::errors::ErrorWrapper;
@@ -12,13 +13,13 @@ use crate::errors::ErrorWrapper;
 #[derive(Clone)]
 #[pyclass(name = "HubHandler")]
 pub struct PyHubHandler {
-    handler: Arc<Mutex<HubHandler>>,
+    handler: Arc<RwLock<HubHandler>>,
 }
 
 impl PyHubHandler {
     pub fn new(handler: HubHandler) -> Self {
         Self {
-            handler: Arc::new(Mutex::new(handler)),
+            handler: Arc::new(RwLock::new(handler)),
         }
     }
 }
@@ -26,20 +27,34 @@ impl PyHubHandler {
 #[pymethods]
 impl PyHubHandler {
     pub async fn refresh_session(&self) -> PyResult<()> {
-        call_handler_method!(self, HubHandler::refresh_session, discard_result)
+        let handler = self.handler.clone();
+        call_handler_method!(
+            handler.write().await.deref_mut(),
+            HubHandler::refresh_session,
+            discard_result
+        )
     }
 
     pub async fn get_device_info(&self) -> PyResult<DeviceInfoHubResult> {
-        call_handler_method!(self, HubHandler::get_device_info)
+        let handler = self.handler.clone();
+        call_handler_method!(handler.read().await.deref(), HubHandler::get_device_info)
     }
 
     pub async fn get_device_info_json(&self) -> PyResult<Py<PyDict>> {
-        let result = call_handler_method!(self, HubHandler::get_device_info_json)?;
+        let handler = self.handler.clone();
+        let result = call_handler_method!(
+            handler.read().await.deref(),
+            HubHandler::get_device_info_json
+        )?;
         Python::with_gil(|py| tapo::python::serde_object_to_py_dict(py, &result))
     }
 
     pub async fn get_child_device_list(&self) -> PyResult<Py<PyList>> {
-        let children = call_handler_method!(self, HubHandler::get_child_device_list)?;
+        let handler = self.handler.clone();
+        let children = call_handler_method!(
+            handler.read().await.deref(),
+            HubHandler::get_child_device_list
+        )?;
 
         let results = Python::with_gil(|py| {
             let results = PyList::empty_bound(py);
@@ -80,12 +95,20 @@ impl PyHubHandler {
     }
 
     pub async fn get_child_device_list_json(&self) -> PyResult<Py<PyDict>> {
-        let result = call_handler_method!(self, HubHandler::get_child_device_list_json)?;
+        let handler = self.handler.clone();
+        let result = call_handler_method!(
+            handler.read().await.deref(),
+            HubHandler::get_child_device_list_json
+        )?;
         Python::with_gil(|py| tapo::python::serde_object_to_py_dict(py, &result))
     }
 
     pub async fn get_child_device_component_list_json(&self) -> PyResult<Py<PyDict>> {
-        let result = call_handler_method!(self, HubHandler::get_child_device_component_list_json)?;
+        let handler = self.handler.clone();
+        let result = call_handler_method!(
+            handler.read().await.deref(),
+            HubHandler::get_child_device_component_list_json
+        )?;
         Python::with_gil(|py| tapo::python::serde_object_to_py_dict(py, &result))
     }
 }